servicio_sensores/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
email-validator
numpy
pandas
scikit-learn
joblib
